+++ src/components/sections/ProductGallery/ProductGallery.tsx
@@ -1,24 +1,23 @@
-import { usePagination, useSearch } from '@faststore/sdk'
+import { usePagination, useSearch as useFSSearch } from '@faststore/sdk'
 import { GatsbySeo } from 'gatsby-plugin-next-seo'
-import { lazy, Suspense, useState } from 'react'
-import Filter from 'src/components/search/Filter'
+import { lazy, useState } from 'react'
 import Sort from 'src/components/search/Sort'
-import FilterSkeleton from 'src/components/skeletons/FilterSkeleton'
 import ProductGridSkeleton from 'src/components/skeletons/ProductGridSkeleton'
 import SkeletonElement from 'src/components/skeletons/SkeletonElement'
 import Button, { ButtonLink } from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
 import { mark } from 'src/sdk/tests/mark'
-import productGalleryQuery from 'src/mocks/productGalleryQuery.json'
-import productGalleryQuery2 from 'src/mocks/productGalleryQuery2.json'
+import useSearch from 'src/contexts/SearchContext/useSearch'
+import Filters from 'src/components/search/PLPFilters'
+import ProductNotFound from 'src/components/common/ProductNotFound/ProductNotFound'
 
 import Section from '../Section'
-import EmptyGallery from './EmptyGallery'
-
+import { useProductsPrefetch } from './usePageProducts'
 import './product-gallery.scss'
+import { useGalleryQuery } from './useGalleryQuery'
+import { useDelayedFacets } from './useDelayedFacets'
 
 const GalleryPage = lazy(() => import('./ProductGalleryPage'))
-const GalleryPageSkeleton = <ProductGridSkeleton loading />
 
 type ForceSvg = {
   svg1?: JSX.Element
@@ -36,36 +35,29 @@ function ProductGallery({
   title,
   searchTerm,
   galleryTitle,
-  forceSvg,
   hasFilter = true,
 }: Props) {
+  const { products, productsCount, isLoading, slug } = useSearch()
+
   const [isFilterOpen, setIsFilterOpen] = useState<boolean>(false)
-  const { pages, addNextPage, addPrevPage, state: searchState } = useSearch()
+  const { pages, addNextPage, addPrevPage, state: searchState } = useFSSearch()
 
-  const { data } = productGalleryQuery // REPLACE QUERY WITH MOCK FILE
+  const { data } = useGalleryQuery()
 
-  // const facets = useDelayedFacets(data)
-
-  const facets = productGalleryQuery.data?.search?.facets // REPLACE QUERY WITH MOCK FILE
-
-  const facetsWithPrice = productGalleryQuery2.data?.search?.facets
-
-  const allFacets = facets.concat(facetsWithPrice)
+  const facets = useDelayedFacets(data)
 
   const totalCount = data?.search.products.pageInfo.totalCount ?? 0
   const { next, prev } = usePagination(totalCount)
 
-  // useProductsPrefetch(prev ? prev.cursor : null)
-  // useProductsPrefetch(next ? next.cursor : null)
+  useProductsPrefetch(prev ? prev.cursor : null)
+  useProductsPrefetch(next ? next.cursor : null)
 
-  if (data && totalCount === 0) {
+  if ((!products || !products.length) && !isLoading) {
     return (
-      <Section
-        data-testid="product-gallery"
-        className="product-listing layout__content"
-      >
-        <EmptyGallery />
-      </Section>
+      <ProductNotFound
+        title="OPS!!!!!"
+        subtitle={`Produtos de "${slug}" não encontrados.`}
+      />
     )
   }
 
@@ -91,24 +83,22 @@ function ProductGallery({
       <div className="product-listing__content-grid layout__content">
         {hasFilter && (
           <div className="product-listing__filters">
-            <FilterSkeleton loading={facets?.length === 0}>
-              <Filter
-                isOpen={isFilterOpen}
-                facets={allFacets}
-                onDismiss={() => setIsFilterOpen(false)}
-                forceSvg={forceSvg && forceSvg}
-              />
-            </FilterSkeleton>
+            <Filters
+              slug={slug}
+              isFilterOpen={isFilterOpen}
+              onDismiss={() => setIsFilterOpen(false)}
+            />
           </div>
         )}
 
-        <div className="product-listing__results-count" data-count={totalCount}>
-          <SkeletonElement shimmer type="text" loading={!data}>
+        <div
+          className="product-listing__results-count"
+          data-count={productsCount}
+        >
+          <SkeletonElement shimmer type="text" loading={!data || isLoading}>
             <h2 data-testid="total-product-count">
-              <span>Mostrando</span>
-              <span>
-                {totalCount} de {totalCount} produtos
-              </span>
+              <span>Monstrando</span>
+              <span>{productsCount} produtos</span>
             </h2>
           </SkeletonElement>
         </div>
@@ -118,7 +108,7 @@ function ProductGallery({
             <Sort />
           </SkeletonElement>
 
-          <SkeletonElement shimmer type="button" loading={facets?.length === 0}>
+          <SkeletonElement shimmer type="button" loading={isLoading}>
             <Button
               data-testid="open-filter-button"
               icon={<Icon name="FadersHorizontal" width={20} height={20} />}
@@ -154,26 +144,27 @@ function ProductGallery({
               </ButtonLink>
             </div>
           )}
-
           {/* Render ALL products */}
-          {data ? (
-            <Suspense fallback={GalleryPageSkeleton}>
-              {pages.map((page) => (
-                <GalleryPage
-                  key={`gallery-page-${page}`}
-                  showSponsoredProducts={false}
-                  fallbackData={page === searchState.page ? data : undefined}
-                  page={page}
-                  title={title}
-                />
-              ))}
-            </Suspense>
-          ) : (
-            GalleryPageSkeleton
-          )}
-
+          <ProductGridSkeleton loading={!products || isLoading}>
+            {pages.map((page) => (
+              <>
+                {products?.length && (
+                  <GalleryPage
+                    key={`gallery-page-${page}`}
+                    showSponsoredProducts={false}
+                    fallbackData={
+                      page === searchState.page ? products : undefined
+                    }
+                    page={page}
+                    title={title}
+                    products={products}
+                  />
+                )}
+              </>
+            ))}
+          </ProductGridSkeleton>
           {/* Add link to next page. This helps on SEO */}
-          {next !== false && (
+          {next !== false && products && products.length ? (
             <div className="product-listing__pagination product-listing__pagination--bottom">
               <GatsbySeo defer linkTags={[{ rel: 'next', href: next.link }]} />
               <ButtonLink
@@ -190,6 +181,8 @@ function ProductGallery({
                 CARREGAR MAIS
               </ButtonLink>
             </div>
+          ) : (
+            <></>
           )}
         </div>
       </div>
@@ -197,5 +190,4 @@ function ProductGallery({
   )
 }
 
-ProductGallery.displayName = 'ProductGallery'
 export default mark(ProductGallery)