+++ src/components/sections/ProductGallery/ProductGalleryPage.tsx
@@ -1,8 +1,7 @@
 import { useSearch } from '@faststore/sdk'
+import ProductNotFound from 'src/components/common/ProductNotFound/ProductNotFound'
+import type { ProductsProductCard } from 'src/components/product/ProductCard/ProductCard'
 import ProductGrid from 'src/components/product/ProductGrid'
-import productGalleryQuery from 'src/mocks/productGalleryQuery.json'
-
-// import { useProducts } from './usePageProducts'
 import './product-gallery.scss'
 
 /* If showSponsoredProducts is true, a ProductTiles will be displayed in between two blocks of ProductGrid on the page 0 */
@@ -11,67 +10,22 @@ interface Props {
   fallbackData?: any
   title: string
   showSponsoredProducts?: boolean
+  products: ProductsProductCard[]
 }
 
-function GalleryPage({
-  page,
-  // fallbackData,
-  showSponsoredProducts = true,
-}: Props) {
-  // const products = useProducts(page, fallbackData)
-
-  const products = productGalleryQuery?.data?.search?.products?.edges
-
+function GalleryPage({ page, products }: Props) {
   const { itemsPerPage } = useSearch()
 
-  if (products == null) {
-    return null
+  if (!products.length) {
+    return (
+      <ProductNotFound
+        title="OPS!!!!!"
+        subtitle="Infelizmente não conseguimos encontrar nenhum resultado para os filtros selecionados."
+      />
+    )
   }
 
-  const productsSponsored = showSponsoredProducts
-    ? products.slice(0, 2)
-    : undefined
-
-  const middleItemIndex = Math.ceil(itemsPerPage / 2)
-
-  const shouldDisplaySponsoredProducts =
-    page === 0 && productsSponsored && productsSponsored.length > 1
-
-  return (
-    <>
-      {/* <Sentinel
-        products={products}
-        page={page}
-        pageSize={itemsPerPage}
-        title={title}
-      /> */}
-      {shouldDisplaySponsoredProducts ? (
-        <>
-          <ProductGrid
-            products={products.slice(0, middleItemIndex)}
-            page={page}
-            pageSize={middleItemIndex}
-          />
-          <div className="product-listing__results-sponsored">
-            <h3>Sponsored</h3>
-            {/*
-              TODO: Refactor this bit of code
-
-              Sections should be self contained and should not import other sections.
-              We should remove/refactor this section from here
-            */}
-          </div>
-          <ProductGrid
-            products={products.slice(middleItemIndex, itemsPerPage)}
-            page={page}
-            pageSize={middleItemIndex}
-          />
-        </>
-      ) : (
-        <ProductGrid products={products} page={page} pageSize={itemsPerPage} />
-      )}
-    </>
-  )
+  return <ProductGrid products={products} page={page} pageSize={itemsPerPage} />
 }
 
 export default GalleryPage