+++ src/components/skeletons/ProductCardSkeleton/ProductCardSkeleton.tsx
@@ -1,6 +1,8 @@
 import Shimmer from '../Shimmer'
 import SkeletonElement from '../SkeletonElement'
 
+import './product-card-skeleton.scss'
+
 interface Props {
   bordered?: boolean
   sectioned?: boolean