+++ src/sdk/product/useProductsQuery.ts
@@ -1,96 +1,134 @@
-// import { useSession } from '@faststore/sdk'
-// import { gql } from '@vtex/graphql-utils'
-// import { useCallback, useMemo } from 'react'
-// import { ITEMS_PER_SECTION } from 'src/constants'
-// import type {
-//   ProductsQueryQuery,
-//   ProductsQueryQueryVariables,
-// } from '@generated/graphql'
+import { useSession } from '@faststore/sdk'
+import { gql } from '@vtex/graphql-utils'
+import { useCallback, useMemo } from 'react'
+import { ITEMS_PER_SECTION } from 'src/constants'
+import type {
+  ProductsQueryQuery,
+  ProductsQueryQueryVariables,
+} from '@generated/graphql'
 
-// import { prefetchQuery } from '../graphql/prefetchQuery'
-// import { useQuery } from '../graphql/useQuery'
-// import type { QueryOptions } from '../graphql/useQuery'
+import { prefetchQuery } from '../graphql/prefetchQuery'
+import { useQuery } from '../graphql/useQuery'
+import type { QueryOptions } from '../graphql/useQuery'
 
-// export const query = gql`
-//   query ProductsQuery(
-//     $first: Int!
-//     $after: String
-//     $sort: StoreSort!
-//     $term: String!
-//     $selectedFacets: [IStoreSelectedFacet!]!
-//   ) {
-//     search(
-//       first: $first
-//       after: $after
-//       sort: $sort
-//       term: $term
-//       selectedFacets: $selectedFacets
-//     ) {
-//       products {
-//         pageInfo {
-//           totalCount
-//         }
-//         edges {
-//           node {
-//             ...ProductSummary_product
-//           }
-//         }
-//       }
-//     }
-//   }
-// `
+export const query = gql`
+  query ProductsQuery(
+    $first: Int!
+    $after: String
+    $sort: StoreSort!
+    $term: String!
+    $selectedFacets: [IStoreSelectedFacet!]!
+  ) {
+    search(
+      first: $first
+      after: $after
+      sort: $sort
+      term: $term
+      selectedFacets: $selectedFacets
+    ) {
+      products {
+        pageInfo {
+          totalCount
+        }
+        edges {
+          node {
+            id: productID
+            sku
+            name
+            gtin
+            description
 
-// const toArray = <T>(x: T[] | T | undefined) =>
-//   Array.isArray(x) ? x : x ? [x] : []
+            isVariantOf {
+              productGroupID
+              name
+            }
 
-// export const useLocalizedVariables = ({
-//   first,
-//   after,
-//   sort,
-//   term,
-//   selectedFacets,
-// }: Partial<ProductsQueryQueryVariables>) => {
-//   const { channel } = useSession()
+            image {
+              url
+              alternateName
+            }
 
-//   return useMemo(() => {
-//     const facets = toArray(selectedFacets)
+            brand {
+              name
+            }
 
-//     return {
-//       first: first ?? ITEMS_PER_SECTION,
-//       after: after ?? '0',
-//       sort: sort ?? ('score_desc' as const),
-//       term: term ?? '',
-//       selectedFacets: [...facets, { key: 'channel', value: channel! }],
-//     }
-//   }, [first, after, sort, term, selectedFacets, channel])
-// }
+            offers {
+              lowPrice
+              offers {
+                availability
+                price
+                listPrice
+                seller {
+                  identifier
+                }
+              }
+            }
 
-// /**
-//  * Use this hook for fetching a list of products, like in search results and shelves
-//  */
-// export const useProductsQuery = (
-//   variables: Partial<ProductsQueryQueryVariables>,
-//   options?: QueryOptions
-// ) => {
-//   const localizedVariables = useLocalizedVariables(variables)
+            breadcrumbList {
+              itemListElement {
+                item
+                name
+                position
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+`
 
-//   const { data } = useQuery<ProductsQueryQuery, ProductsQueryQueryVariables>(
-//     query,
-//     localizedVariables,
-//     options
-//   )
+const toArray = <T>(x: T[] | T | undefined) =>
+  Array.isArray(x) ? x : x ? [x] : []
 
-//   return data?.search.products
-// }
+export const useLocalizedVariables = ({
+  first,
+  after,
+  sort,
+  term,
+  selectedFacets,
+}: Partial<ProductsQueryQueryVariables>) => {
+  const { channel } = useSession()
 
-// export const useProductsQueryPrefetch = (
-//   variables: ProductsQueryQueryVariables,
-//   options?: QueryOptions
-// ) => {
-//   const localizedVariables = useLocalizedVariables(variables)
+  return useMemo(() => {
+    const facets = toArray(selectedFacets)
 
-//   return useCallback(
-//     () => prefetchQuery(query, localizedVariables, options),
-//     [options, localizedVariables]
-//   )
-// }
+    return {
+      first: first ?? ITEMS_PER_SECTION,
+      after: after ?? '0',
+      sort: sort ?? ('score_desc' as const),
+      term: term ?? '',
+      selectedFacets: [...facets, { key: 'channel', value: channel! }],
+    }
+  }, [first, after, sort, term, selectedFacets, channel])
+}
+
+/**
+ * Use this hook for fetching a list of products, like in search results and shelves
+ */
+export const useProductsQuery = (
+  variables: Partial<ProductsQueryQueryVariables>,
+  options?: QueryOptions
+) => {
+  const localizedVariables = useLocalizedVariables(variables)
+
+  const { data } = useQuery<ProductsQueryQuery, ProductsQueryQueryVariables>(
+    query,
+    localizedVariables,
+    options
+  )
+
+  return data?.search.products
+}
+
+export const useProductsQueryPrefetch = (
+  variables: ProductsQueryQueryVariables,
+  options?: QueryOptions
+) => {
+  const localizedVariables = useLocalizedVariables(variables)
+
+  return useCallback(
+    () => prefetchQuery(query, localizedVariables, options),
+    [options, localizedVariables]
+  )
+}