+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]-
 
+## [0.14.0] - 2022-08-24
+
+### Added
+
+- **GM-343** - Make Department/Category pages dinamic
+- **GM-344** - Integrate filters for Department/Category page
+
 ## [0.13.0] - 2022-08-18
 
 ### Added